+++ pytemscript/utils/constants.py
@@ -1,6 +1,6 @@
 SCRIPTING_TECNAI = "Tecnai.Instrument"
 SCRIPTING_STD = "TEMScripting.Instrument.1"
-SCRIPTING_ADV = "TEMAdvancedScripting.AdvancedInstrument.2"
+SCRIPTING_ADV = "TEMAdvancedScripting.AdvancedInstrument.2"  # .1 for Titan TEM server < 2.15 ?
 SCRIPTING_LOWDOSE = "LDServer.LdSrv"
 SCRIPTING_TIA = "ESVision.Application"
 SCRIPTING_SEM_CCD = "SerialEMCCD.DMCamera"